+++ src/multi-select-item/index.js
@@ -1,0 +1,1 @@
+export MultiSelectItem from './multi-select-item.component.jsx';