+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
+## <next patch>
+
+* `ApplicationLayout` disabled touch for sidebar: drag area on left is reserved by browser e.g. in iPad, sidebar burger icon fill is white.
+* npm script `examples` is now `docs` and related files are renamed accordingly
+
 ## 2.0.0
 
 * Change package name from 'ocfrontend' to 'oc-common-ui'